chore(): revert creation of new pages
