+++ extensions/apps/antenna/src/components/pages/entry-page/reflection-page.tsx
@@ -3,7 +3,7 @@ import Card from '@akashaorg/design-system-core/lib/components/Card';
 import Stack from '@akashaorg/design-system-core/lib/components/Stack';
 import BackToOriginalBeam from '@akashaorg/ui-lib-feed/lib/components/back-to-original-beam';
 import ReflectionSection from './reflection-section';
-import ReflectFeed from '@akashaorg/ui-lib-feed/lib/components/reflect-feed';
+import ReflectFeed from '@akashaorg/ui-lib-feed/lib/components/reflect-feed.new';
 import Divider from '@akashaorg/design-system-core/lib/components/Divider';
 import ErrorBoundary from '@akashaorg/design-system-core/lib/components/ErrorBoundary';
 import {
@@ -14,11 +14,11 @@ import {
   useRootComponentProps,
 } from '@akashaorg/ui-awf-hooks';
 import { useTranslation } from 'react-i18next';
-import { EntityTypes } from '@akashaorg/typings/lib/ui';
 import { ReflectionPreview } from '@akashaorg/ui-lib-feed';
 import { useNavigate } from '@tanstack/react-router';
 import { GetReflectionByIdQuery } from '@akashaorg/typings/lib/sdk/graphql-operation-types-new';
 import { EditableReflectionResolver } from '@akashaorg/ui-lib-feed/lib/components/editable-reflection/editable-reflection-resolver';
+import { EntityTypes } from '@akashaorg/typings/lib/ui';
 
 type ReflectionPageProps = {
   reflection: GetReflectionByIdQuery;
@@ -31,7 +31,7 @@ const ReflectionPage: React.FC<ReflectionPageProps> = props => {
     data: { authenticatedDID },
   } = useAkashaStore();
   const isLoggedIn = !!authenticatedDID;
-  const { navigateToModal, getTranslationPlugin, logger } = useRootComponentProps();
+  const { navigateToModal, logger } = useRootComponentProps();
   const [analyticsActions] = useAnalytics();
   const wrapperRef = useRef(null);
   const navigate = useNavigate();
@@ -63,32 +63,27 @@ const ReflectionPage: React.FC<ReflectionPageProps> = props => {
   return (
     <Card padding="p-0" margin="mb-4">
       <Stack spacing="gap-y-2">
+        <>
+          <BackToOriginalBeam
+            label={t('Back to original beam')}
+            onClick={() => onNavigateToOriginalBeam(entryData.beam?.id)}
+          />
+          <>
+            <ReflectionSection
+              beamId={entryData.beam?.id}
+              reflectionId={entryData.id}
+              entryData={mapReflectEntryData(entryData)}
+              isLoggedIn={isLoggedIn}
+              parentWrapperRef={wrapperRef}
+              showLoginModal={showLoginModal}
+            />
+          </>
+        </>
         <ReflectFeed
-          header={
-            <>
-              <BackToOriginalBeam
-                label={t('Back to original beam')}
-                onClick={() => onNavigateToOriginalBeam(entryData.beam?.id)}
-              />
-              <>
-                <ReflectionSection
-                  beamId={entryData.beam?.id}
-                  reflectionId={entryData.id}
-                  entryData={mapReflectEntryData(entryData)}
-                  isLoggedIn={isLoggedIn}
-                  parentWrapperRef={wrapperRef}
-                  showLoginModal={showLoginModal}
-                />
-              </>
-            </>
-          }
-          reflectionsOf={{ entryId: entryData.id, itemType: EntityTypes.REFLECT }}
+          itemType={EntityTypes.REFLECT}
           itemSpacing={0}
-          newItemsPublishedLabel={t('New Reflects published recently')}
           trackEvent={analyticsActions.trackEvent}
-          locale={getTranslationPlugin().i18n.language}
           estimatedHeight={120}
-          queryKey={`reflect-feed-${entryData.id}`}
           filters={{
             and: [
               { where: { isReply: { equalTo: true } } },
@@ -112,12 +107,12 @@ const ReflectionPage: React.FC<ReflectionPageProps> = props => {
               <>
                 <Divider />
                 <EditableReflectionResolver
-                  reflectID={itemData.node.reflectionID}
+                  reflectID={itemData.reflectionID}
                   onContentClick={() => {
                     navigate({
                       to: '/reflection/$reflectionId',
                       params: {
-                        reflectionId: itemData.node.reflectionID,
+                        reflectionId: itemData.reflectionID,
                       },
                     });
                   }}
@@ -125,13 +120,13 @@ const ReflectionPage: React.FC<ReflectionPageProps> = props => {
                     navigate({
                       to: '/reflection/$reflectionId/reflect',
                       params: {
-                        reflectionId: itemData.node.reflectionID,
+                        reflectionId: itemData.reflectionID,
                       },
                     });
                   }}
                 />
                 <ReflectionPreview
-                  reflectionId={itemData.node.id}
+                  reflectionId={itemData.id}
                   onNavigate={(options: { id: string; reflect?: boolean }) => {
                     navigate({
                       to: options.reflect
